Efficiency improvement - package search results are cached.
Whenever a package entry is located, it is stored in a cache, which is a subset
of the package list it was found in. This cache is used in subsequent package
searches, to improve efficiency. Since the same selection of packages is used
across roar-ng's scripts, this results in a major boost to roar-ng's overall
performance.

In a typical roar-ng session, once 1download is called, any other script which
searches for package entries (e.g a second call to 1download or 2createpackages)
is much more efficient.