Bug fix - /usr/bin/python3 is a broken symlink.
Under Debian (and other distributions), /usr/bin/python is a symlink to an
executable, which contains the Python version it corresponds to in its name.
However, some packages rely on /usr/bin/python3, a symlink to the default Python
3 interpreter.

This symlink was broken, since Debian's Python 3 package contains multiple files
which match "python3*" under /usr/bin.