Implemented support for per-repository mirrors.
roar-ng now supports different mirrors for the various package repositories of a
given distribution. This support is required in order to support partial mirrors
(which mirror only a given repository) or package lists with different base
paths, under the same mirror.

In addition, get_package_entry() now appends a field to its output package
entry, which contains the name of the repository it was found in. This field is
used to find download mirrors for the package.