Added support for per-distribution package list download.
This support is required for RPM-based distributions, since the package list
file name varies.

From now and on, instead of downloading the file specified as the package list
URL, an external script is responsible for downloading the package list,
according to the file pointed by this URL. In the case of CentOS, it points
to the repository base directory; the script downloads repomd.xml, filters the
package list file name and downloads it.